+++ src/applications/edu-benefits/10203/config/chapters.js
@@ -2,13 +2,14 @@ import fullSchema10203 from 'vets-json-schema/dist/22-10203-schema.json';
 import createContactInformationPage from '../../pages/contactInformation';
 import createApplicantInformationPage from 'platform/forms/pages/applicantInformation';
 
-import { display10203StemFlow } from '../helpers';
+import { display10203StemFlow, displayConfirmEligibility } from '../helpers';
 
 import {
   activeDuty,
   benefitSelection,
   stemEligibility,
   confirmEligibility,
+  programDetails,
 } from '../pages';
 
 export const chapters = {
@@ -37,7 +38,7 @@ export const chapters = {
     title: 'Program Details',
     pages: {
       stemEligibility: {
-        title: 'Rogers STEM Scholarship eligibility',
+        title: 'STEM Scholarship eligibility',
         path: 'benefits/stem-eligibility',
         uiSchema: stemEligibility.uiSchema,
         schema: stemEligibility.schema,
@@ -45,9 +46,16 @@ export const chapters = {
       confirmEligibility: {
         title: '',
         path: 'benefits/confirm-eligibility',
+        depends: form => displayConfirmEligibility(form),
         pageClass: 'vads-u-max-width--100 vads-u-vads-u-width--full',
         uiSchema: confirmEligibility.uiSchema,
         schema: confirmEligibility.schema,
+      },
+      programDetails: {
+        title: 'Your STEM degree',
+        path: 'benefits/program-details',
+        uiSchema: programDetails.uiSchema,
+        schema: programDetails.schema,
       },
     },
   },
