+++ src/applications/edu-benefits/10203/config/form.js
@@ -1,4 +1,4 @@
-import fullSchema10203 from 'vets-json-schema/dist/22-10203-schema.json';
+import fullSchema from 'vets-json-schema/dist/22-10203-schema.json';
 
 import { transform } from '../submit-transformer';
 import { prefillTransformer } from '../prefill-transformer';
@@ -16,13 +16,6 @@ import preSubmitInfo from 'platform/forms/preSubmitInfo';
 import IntroductionPage from '../containers/IntroductionPage';
 import ConfirmationPage from '../containers/ConfirmationPage';
 import { chapters } from './chapters';
-
-const {
-  preferredContactMethod,
-  date,
-  dateRange,
-  serviceBefore1977,
-} = fullSchema10203.definitions;
 
 const formConfig = {
   urlPrefix: '/',
@@ -43,10 +36,7 @@ const formConfig = {
   introduction: IntroductionPage,
   confirmation: ConfirmationPage,
   defaultDefinitions: {
-    preferredContactMethod,
-    serviceBefore1977,
-    date,
-    dateRange,
+    ...fullSchema.definitions,
   },
   title: 'Update your education benefits',
   subTitle: 'Form 22-10203',
