+++ src/applications/edu-benefits/10203/content/programDetails.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+const schoolStudentIdTitle = (
+  <div>
+    <p>
+      <strong>Your school ID and contact details</strong>
+    </p>
+    <p>
+      These details will help us review your application faster, but aren't
+      required.
+    </p>
+  </div>
+);
+
+export { schoolStudentIdTitle };
