+++ src/applications/edu-benefits/10203/content/stemEligibility.jsx
@@ -6,7 +6,7 @@ const benefitLeftDescription = (
     target="_blank"
     rel="noopener noreferrer"
   >
-    Check remaining benefits
+    Check your remaining benefits
   </a>
 );
