+++ src/applications/edu-benefits/10203/pages/index.js
@@ -1,6 +1,13 @@
 import * as activeDuty from './activeDuty';
 import * as benefitSelection from './benefitSelection';
-import * as stemEligibility from './stemEligibility';
 import * as confirmEligibility from './confirmEligibility';
+import * as programDetails from './programDetails';
+import * as stemEligibility from './stemEligibility';
 
-export { activeDuty, benefitSelection, stemEligibility, confirmEligibility };
+export {
+  activeDuty,
+  benefitSelection,
+  stemEligibility,
+  confirmEligibility,
+  programDetails,
+};
